requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiosqlite>=0.19.0
